+++ src/main/java/com/project/numble/application/bookmark/controller/BookmarkController.java
@@ -1,6 +1,6 @@
-package com.project.numble.application.board.controller;
+package com.project.numble.application.bookmark.controller;
 
-import com.project.numble.application.board.service.LikeService;
+import com.project.numble.application.bookmark.service.BookmarkService;
 import com.project.numble.core.resolver.SignInUser;
 import com.project.numble.core.resolver.UserInfo;
 import lombok.RequiredArgsConstructor;
@@ -10,23 +10,20 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/like")
-public class LikeController {
+@RequestMapping("/bookmark")
+public class BookmarkController {
 
-    private final LikeService likeService;
+    private final BookmarkService bookmarkService;
 
     @PostMapping("/{boardId}")
-    public ResponseEntity<Void> addLike(@PathVariable("boardId") Long boardId, @SignInUser UserInfo userInfo) {
-
-        likeService.addLike(userInfo.getUserId(), boardId);
-
+    public ResponseEntity<Void> addBookmark(@PathVariable("boardId") Long boardId, @SignInUser UserInfo userInfo) {
+        bookmarkService.addBookmark(userInfo.getUserId(), boardId);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{boardId}")
     public ResponseEntity<Void> cancelLike(@PathVariable("boardId") Long boardId, @SignInUser UserInfo userInfo) {
-        likeService.cancelLike(userInfo.getUserId(), boardId);
+        bookmarkService.cancelBookmark(userInfo.getUserId(), boardId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
-
 }
