+++ src/main/java/com/project/numble/application/bookmark/domain/Bookmark.java
@@ -1,5 +1,6 @@
-package com.project.numble.application.board.domain;
+package com.project.numble.application.bookmark.domain;
 
+import com.project.numble.application.board.domain.Board;
 import com.project.numble.application.user.domain.User;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
@@ -10,11 +11,10 @@ import javax.persistence.*;
 
 @Entity
 @Getter
-@Table(name = "tb_likes")
+@Table(name = "tb_bookmark")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EqualsAndHashCode(of = "id", callSuper = false)
-public class Like {
-
+public class Bookmark {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -25,10 +25,11 @@ public class Like {
     @ManyToOne(fetch = FetchType.LAZY)
     private Board board;
 
-    public Like(User user, Board board) {
+    public Bookmark(User user, Board board) {
         this.user = user;
         this.board = board;
     }
+
     public void initBoard(Board board) {
         if (this.board == null) {
             this.board = board;
