+++ src/main/java/com/project/numble/application/bookmark/service/exception/AlreadyBookmarkBoardException.java
@@ -1,4 +1,4 @@
-package com.project.numble.application.board.service.exception;
+package com.project.numble.application.bookmark.service.exception;
 
 public class AlreadyBookmarkBoardException extends RuntimeException{
 }
