+++ src/main/java/com/project/numble/application/like/controller/LikeController.java
@@ -1,6 +1,6 @@
-package com.project.numble.application.board.controller;
+package com.project.numble.application.like.controller;
 
-import com.project.numble.application.board.service.BookmarkService;
+import com.project.numble.application.like.service.LikeService;
 import com.project.numble.core.resolver.SignInUser;
 import com.project.numble.core.resolver.UserInfo;
 import lombok.RequiredArgsConstructor;
@@ -10,20 +10,23 @@ import org.springframework.web.bind.annotation.*;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/bookmark")
-public class BookmarkController {
+@RequestMapping("/like")
+public class LikeController {
 
-    private final BookmarkService bookmarkService;
+    private final LikeService likeService;
 
     @PostMapping("/{boardId}")
-    public ResponseEntity<Void> addBookmark(@PathVariable("boardId") Long boardId, @SignInUser UserInfo userInfo) {
-        bookmarkService.addBookmark(userInfo.getUserId(), boardId);
+    public ResponseEntity<Void> addLike(@PathVariable("boardId") Long boardId, @SignInUser UserInfo userInfo) {
+
+        likeService.addLike(userInfo.getUserId(), boardId);
+
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @DeleteMapping("/{boardId}")
     public ResponseEntity<Void> cancelLike(@PathVariable("boardId") Long boardId, @SignInUser UserInfo userInfo) {
-        bookmarkService.cancelBookmark(userInfo.getUserId(), boardId);
+        likeService.cancelLike(userInfo.getUserId(), boardId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
+
 }
