+++ src/main/java/com/project/numble/application/like/domain/Like.java
@@ -1,5 +1,6 @@
-package com.project.numble.application.board.domain;
+package com.project.numble.application.like.domain;
 
+import com.project.numble.application.board.domain.Board;
 import com.project.numble.application.user.domain.User;
 import lombok.AccessLevel;
 import lombok.EqualsAndHashCode;
@@ -10,10 +11,11 @@ import javax.persistence.*;
 
 @Entity
 @Getter
-@Table(name = "tb_bookmark")
+@Table(name = "tb_likes")
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @EqualsAndHashCode(of = "id", callSuper = false)
-public class Bookmark {
+public class Like {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -24,11 +26,10 @@ public class Bookmark {
     @ManyToOne(fetch = FetchType.LAZY)
     private Board board;
 
-    public Bookmark(User user, Board board) {
+    public Like(User user, Board board) {
         this.user = user;
         this.board = board;
     }
-
     public void initBoard(Board board) {
         if (this.board == null) {
             this.board = board;
